+++ src/squidpy/_settings/__init__.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from squidpy.settings._dispatch import gpu_dispatch
-from squidpy.settings._settings import DeviceType, settings
+from squidpy._settings._dispatch import gpu_dispatch
+from squidpy._settings._settings import DeviceType, settings
 
 __all__ = ["settings", "DeviceType", "gpu_dispatch"]
