+++ src/squidpy/_settings/_settings.py
